[ARA] Add more support for ARA analysis progress report
Additional listeners other than the host can now be reported to.

Unlike in the other notification methods, we delay the reporting of the analysis progress until we are in the proper notification thread (usually the main thread). The reason we do this is that the most common use-case is probably to update some GUI component and repaint it, and calling it to repaint must be done on the main thread.