Avoid sending extraneous notifications to scrollbars listeners
Following all code paths it seems this notifications aren't actually needed, and they -can- trigger undesired behavior on startup (depending on what the listeners do).
The intention here is to have the listeners only get a notification when the user actualy moves the view via the scrollbars, not in other cases.