Updating ARASecondsPixelMapper to allocate its tempo / bar signatures content event readers on the stack.
This was done in order to prevent multiple HostContentReaders of the same content existing simultaneously - readers are meant to be stack objects on the ARA model thread, and there shouldn't be a situation that requires more than one for the same data. The resulting code is a bit repetitive, though - could we make a new converter class that manages a HostContentReader?