Proposed fix for soundradix/pajam#704, fullscreen switch crash
My first fix invloved correctly retaining/releasing the colourSpace
member whenever it was changed, so it was never released while we still
had a reference to it.

But then I realized that basically we're always just updating it to be
equal to [[[view window] screen] colorSpace], and it's better not to
have to maintain state if you can.

This does make one change which I don't understand the significance of
-- in paint() we are now always calling updateViewportSize with
canTriggerUpdate = true, instead of only passing canTriggerUpdate when
the colorspace changed.  But I don't see why, if we were allowed to
update when the colorspace changed, why we wouldn't always be allowed to
update.  I don't think there should be a huge performance impact because
updateViewportSize only invalidates when the viewport actually changes.
Please advice.