Revert "fix for performance in JUCE-based ARA hosts"
This change is no longer present in either the JUCE develop branch or in Celemony's JUCE_ARA fork.
I noticed this because of other JUCE changes which broke this code.
Instead of maintaining, I'll just revert this remainder which seems to only be currently present on our branch. It should probably have been removed anyway in our merges when it was removed from Celemony's fork.
I don't know if this would regress ARA hosts performance as originally claimed. We can discuss further with Celemony to figure out the up-to-date status.

This reverts commits 983195684c3842d31bc50642e71acddbbad8b801 and 8c7b6e6ddc373cdb04f5a1ba32ff6c8c2270374e.